Refactor: inline analytics as pageBlock below store list table
